requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Session==0.5.0
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
Pillow==10.0.0
scikit-learn==1.3.0
tensorflow==2.13.0
Werkzeug==2.3.7
gunicorn==21.2.0
